Multiple changes:
1)  Add Dockerfile.AzureLinux as another example

This Dockerfile also has the same variants as the original but
is based on the official Microsoft DotNet SDK on Azure Linux 3.0 image.
This shows how much they are the same and yet different.

2)  Clean up variants

We put make and gdb into the base image now such that we don't
need to add it everywhere.  Both are useful in most any development
work stream.

3)  Update the README to talk about the new Dockerfile

4)  Add a bit about building your own variants

5)  Add a bit in security about the isolation keeping the user safe