Changed to a better logging of the effective command line
This does a better job of making the effective command line that is
logged to be really cut-n-paste ready (and even scriptable)

Most of the time it is not a problem but when certain characters are in
parameters, it may require special escaping and quoting to accurately
represent the value.  We actually use the bash command logging mechanism
to log the command that we were started with (effectively, via all
argument values) and capture that and clean it up such that we can
get show it.