Now with tested Windows support!
Well, you need Docker Desktop for Windows installed (and running)
and you need the Windows GIT installed but then this works with
the few minor tweaks.  Both the build.cmd and boxed.cmd work
and boxed.cmd (plus its boxed.sh) can be put anywhere on the path
and it will work as it does in Linux and MacOS!

Time to let you CycoD loose to play on your Windows.

The next trick will be to build something like this with Windows
containers (do we even need that?)