Update parser to not depend on column command
While this command is old and has been on every Unix, FreeBSD, and Linux
system I have ever worked on, it seems that sometimes, in minimal
systems, it is not included.  (It is included in busybox...)

Anyway, with a bit of work we can just use printf to deal with it.
It works well and is a 100% built-in.

(Note column even existed in the bash environment within git!)