Main change - auto-detect CPU arch for Arm64 vs Amd64
To help make things a bit better for Arm64 and Amd64 machines,
the scripts set the default platform to the one that match the
running environment if they are one of the two.  The fallback is
that it defaults to amd64 (Intel x86_64) just as before.

I made some minor changes to the Dockerfile to make sure the
file tool is always available (always useful) and that we have
the doxygen code documentation creation tool available in case we
want to use it.

Updated the README to explain a bit about the (limited) auto-detect
for the default platform value.

Changed the image targets for the Dockerfile.AzureLinux so that I
can have both built at the same time without a different image name.